Split off phaser defaults tests. Add test for merging properties.
