+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "genie",
-    "version": "3.4.2",
+    "version": "3.4.3",
     "description": "Genie - A Modular Game Engine",
     "license": "Apache-2.0",
     "private": true,
