+++ src/core/loader/phaserDefaults.js
@@ -26,7 +26,7 @@ export const getPhaserDefaults = config => {
         ...{
             width: CANVAS_WIDTH,
             height: CANVAS_HEIGHT,
-            renderer: browser.forceCanvas ? Phaser.CANVAS : Phaser.AUTO,
+            type: browser.forceCanvas ? Phaser.CANVAS : Phaser.AUTO,
             antialias: true,
             multiTexture: true,
             parent: getContainerDiv(),
